practicals/model_deployment/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
enum-compat==0.0.3
future==0.18.2
idna==3.4
imageio==2.22.4
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
psutil==5.9.4
pyparsing==3.0.9
requests==2.28.1
torch==1.13.0
torch-model-archiver==0.4.2
torch-workflow-archiver==0.1.2
torchserve==0.5.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.12
